// HelpDeskFE/HelpDeskFETp/src/api/model/createTypeDevicesDto.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * HelpDesk API
 * OpenAPI spec version: v1
 */

export interface CreateTypeDevicesDto {
  /**
     * @minLength 0
     * @maxLength 60
     */
  name: string;
  /**
     * @minLength 0
     * @maxLength 240
     * @nullable
     */
  description?: string | null;
}